Completed search for single card and displaying it to text area

+++ com.cn55.program/src/cn55/view/MainFrame.java
@@ -1,6 +1,7 @@
 package cn55.view;
 
 import cn55.model.Shop;
+import jdk.nashorn.internal.scripts.JO;
 
 import javax.swing.*;
 import java.awt.*;
@@ -8,7 +9,8 @@ import java.util.HashMap;
 
 public class MainFrame extends JFrame {
 
-    private JTabbedPane tabbedPane;
+    private Shop shop;
+    private JTabbedPane tabPane;
     //private JPanel welcomePanel;
     private CardPanel cardPanel;
     private PurchasesPanel purchasesPanel;
@@ -18,11 +20,13 @@ public class MainFrame extends JFrame {
     public MainFrame(Shop shop) {
         super("Marvel Rewards");
 
+        this.shop = shop;
+
         setLayout(new BorderLayout());
 
         //mainToolbar = new MainToolbar();
         this.tabPane = new JTabbedPane();
-        tabPane.setFont(new Font("Verdana", Font.BOLD, 32));
+
 
         /*// Initialize panels for tabs
         this.welcomePanel = new JPanel();
@@ -41,6 +45,10 @@ public class MainFrame extends JFrame {
 
         // Add panels, toolbars, and panes to main Frame
         //add(mainToolbar, BorderLayout.NORTH);
+        tabPane.setBackground(Style.red500());
+        tabPane.setForeground(Style.btnTextColor());
+        tabPane.setFont(Style.tabPaneFont());
+
         add(tabPane, BorderLayout.CENTER);
 
         // Add tabs to tabPane group
@@ -66,34 +74,50 @@ public class MainFrame extends JFrame {
         tabPane.addChangeListener(e -> {
             if (tabPane.getSelectedComponent() == cardPanel) {
                 cardPanel.refresh();
-            } else {
-                // TODO DO NOT LEAVE THIS LIKE THIS - USED FOR TESTING ONLY
-                cardPanel.refresh();
             }
         });
 
         // Listens to events in Toolbar and handles the event
         cardPanel.setCreateCardListener(new FormListener() {
-            public void cardFormActionEvent() {
+            public void cardFormActionOccurred() {
                 CardForm createCardForm = new CardForm(shop.generateCardID());
                 HashMap<String,String> newCard = createCardForm.getCardMap();
                 shop.makeCard(newCard);
+                cardPanel.refresh();
             }
         });
 
         cardPanel.setDeleteCardListener(new FormListener() {
-            public void cardFormActionEvent() {
+            public void cardFormActionOccurred() {
                 CardForm deleteCardForm = new CardForm();
 
-                if (!shop.cardExists(deleteCardForm.getCardID())) {
+                if (deleteCardForm.getCardID() != null
+                        && shop.cardExists(deleteCardForm.getCardID())) {
                     System.out.println("MainFrame");
                     System.out.println(deleteCardForm.getCardID());
                     deleteCardForm.deleteForm(false);
                 } else {
                     System.out.println("Deleting Card: " + deleteCardForm.getCardID());
                     shop.deleteCard(deleteCardForm.getCardID());
+                    cardPanel.refresh();
                 }
+            }
+        });
 
+        cardPanel.setSearchListener(new SearchListener() {
+            public void searchEventOccurred(SearchEvent e) {
+                String cardID = e.getSearchID();
+
+                if (!cardID.isEmpty() && shop.cardExists(cardID)) {
+                    int cardIndex = shop.getDatabase().getCardMap().get(cardID);
+                    String cardText = shop.getDatabase().getCards().get(cardIndex).toString();
+                    cardPanel.appendCardTextArea(cardText);
+                } else if (!shop.cardExists(cardID)) {
+                    JOptionPane.showMessageDialog(null,
+                            "Card Does Not Exist",
+                            "Error",
+                            JOptionPane.ERROR_MESSAGE);
+                }
             }
         });
     }
